requirements.txt
Flask==3.0.3
gspread==6.0.2
google-auth==2.23.4
requests==2.31.0
beautifulsoup4==4.12.3
gunicorn==21.2.0